Persist filter/search via url
Closes #77
